refactor: move commands to actor companion object
